refactors into chat wrapper
